+++ component/backend/View/Items/Html.php
@@ -64,19 +64,7 @@ class Html extends BaseView
 		{
 			$function = $this->input->getCmd('function', 'arsSelectItem');
 
-			$js = <<< JS
-function arsItemsProxy(id, title)
-{
-    if (window.parent)
-    {
-        window.parent.{$function}(id, title);
-    }
-}
-
-JS;
-
-			$this->addJavascriptInline($js);
-
+			$this->getContainer()->platform->addScriptOptions('ars.itemsProxyCallback', $function);
 		}
 	}